requirements.txt
pandas
requests
python-dotenv
SQLAlchemy
psycopg2-binary

